fix stop script

+++ scripts/stop_app.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 
 PID=$(ps -C node -o pid=)
-kill -9 "$PID"
+if [ ! "$PID" = "" ] ; then kill -9 "$PID" ; fi
